requirements.txt
amqp==5.2.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
billiard==4.2.1
bleach==6.1.0
bootstrap4==0.1.0
Brotli==1.1.0
celery==5.4.0
certifi==2022.12.7
cffi==1.17.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cron-descriptor==1.4.5
cssselect2==0.8.0
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
dj-database-url==2.3.0
Django==5.1.1
django-celery-beat==2.7.0
django-cloudinary-storage==0.3.0
django-crispy-forms==2.3
django-heroku==0.3.1
django-timezone-field==7.0
django-widget-tweaks==1.5.0
docutils==0.20.1
et_xmlfile==2.0.0
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.18.0
fonttools==4.53.1
fqdn==1.5.1
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.2
ibm-cloud-sdk-core==3.20.3
ibm-watson==8.1.0
idna==3.4
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
Kivy==2.2.0
kivy-deps.angle==0.3.3
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.6.0
Kivy-Garden==0.1.5
kivymd==1.1.1
kiwisolver==1.4.5
kombu==5.4.2
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mistune==3.0.2
mysql-connector-python==8.0.33
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.2.2
notebook_shim==0.2.4
numpy==2.0.0
openpyxl==3.1.5
overrides==7.7.0
packaging==23.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pillow==11.3.0
pipenv==2024.4.1
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==3.20.3
psutil==6.0.0
psycopg2==2.9.10
pure_eval==0.2.3
pycparser==2.22
pydyf==0.11.0
Pygments==2.15.1
PyJWT==2.8.0
pyparsing==3.1.2
pyphen==0.17.2
pypiwin32==223
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-json-logger==2.0.7
pytimeparse2==1.7.1
pytz==2024.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.2
pyzmq==26.2.0
redis==5.2.0
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.0
scikit-learn==1.5.1
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
simplejson==3.19.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlite3-to-mysql==2.0.0
sqlparse==0.5.1
stack-data==0.6.3
tabulate==0.9.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tornado==6.4.1
tqdm==4.65.0
traitlets==5.14.3
types-python-dateutil==2.9.0.20240821
typing_extensions==4.6.3
tzdata==2024.1
Unidecode==1.3.6
uri-template==1.3.0
urllib3==2.2.2
validate_email==1.3
vboxapi==1.0
vine==5.1.0
virtualenv==20.30.0
wcwidth==0.2.13
weasyprint==53.3
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
whitenoise==6.9.0
widgetsnbextension==4.0.13
zopfli==0.2.3.post1
